[update] fix name_kana policy
