+++ src/details/line/99112.json
@@ -1,5 +1,5 @@
 {
-  "name":"天塩炭鉱鉄道",
+  "name":"天塩炭礦鉄道",
   "station_list":[
     {"code":1111312,"name":"留萌"},
     {"code":9911202,"name":"春日町(北海道)"},
