+++ src/line.csv
@@ -654,7 +654,7 @@ code,id,name,name_kana,name_formal,station_size,company_code,color,symbol,closed
 99109,c2a803,ふるさと銀河線,ふるさとぎんがせん,北海道ちほく高原鉄道ふるさと銀河線,33,307,NULL,NULL,1,2006-04-21,0
 99110,9ca86d,定山渓鉄道線,じょうざんけいてつどうせん,定山渓鉄道線,20,NULL,NULL,NULL,1,1969-11-01,0
 99111,4d763a,寿都鉄道,すっつてつどう,寿都鉄道,5,NULL,NULL,NULL,1,1972-05-11,0
-99112,032d31,天塩炭鉱鉄道,てしおたんこうてつどう,天塩炭鉱鉄道,8,NULL,NULL,NULL,1,1967-07-31,0
+99112,032d31,天塩炭礦鉄道,てしおたんこうてつどう,天塩炭礦鉄道,8,NULL,NULL,NULL,1,1967-07-31,0
 99113,e0abc3,羽幌炭礦鉄道,はぼろたんこうてつどう,羽幌炭礦鉄道,8,NULL,NULL,NULL,1,1970-12-15,0
 99114,323c43,北海道拓殖鉄道,ほっかいどうたくしょくてつどう,北海道拓殖鉄道,15,NULL,NULL,NULL,1,1968-10-01,0
 99115,d339e1,夕張鉄道線,ゆうばりてつどうせん,北海道炭礦汽船夕張鉄道線,22,NULL,NULL,NULL,1,1975-04-01,0
