+++ src/manual.md
@@ -28,7 +28,7 @@
 - solved/line.json 駅メモ路線一覧
 
 ```
-$ ruby script/check.rb
+$ ruby src/script/check.rb
 ```
 
 ### 確認項目
@@ -43,19 +43,20 @@ $ ruby script/check.rb
 
 ## 駅のボロノイ図・Kd-tree
 
-`solved/station.json`を入力としてjavaで計算  
+`src/solved/station.json`を入力としてjavaで計算  
 
 ```
-script/diagram.bat
+src/script/diagram.bat
 ```
 
-`diagram/station.json`に出力される
+`src/diagram/station.json`に出力される
 
 ## 路線のポリラインデータ
 
 `solved/line.json`を基に`polyline/raw/*.json`=>`polyline/solved/*.json`に出力する
 
 ```
+cd src
 del polyline/solved/*
 script/polyline.bat
 ```
@@ -69,10 +70,10 @@ script/polyline.bat
 - details/line/*.json 路線詳細
 - polyline/solved/*.json 路線ポリライン
 ```
-ruby script/process.rb [version]
+ruby src/script/process.rb [version]
 ```
 
-`../out`に出力
+`./out`に出力
 
 ## GitHub
