+++ src/script/check.rb
@@ -1,7 +1,6 @@
-load("script/utils.rb")
+load("src/script/format.rb")
 
 require "net/http"
-require "minitest/autorun"
 
 STATION_FIELD = [
   "code",
@@ -36,12 +35,12 @@ LINE_FIELD = [
   "impl",
 ]
 
-API_KEY = read("api_key.txt")
+API_KEY = read("src/api_key.txt")
 
 def get_address(station)
   print "get address of station:#{station["name"]} > "
   data = nil
-  file = "details/address/#{station["code"]}.json"
+  file = "src/details/address/#{station["code"]}.json"
   if !data
     uri = URI.parse("https://maps.googleapis.com/")
     https = Net::HTTP.new(uri.host, uri.port)
@@ -58,7 +57,7 @@ def get_address(station)
     h["lng"] = station["lng"]
     h["plus_code"] = data["plus_code"]
     h["results"] = data["results"]
-    File.open("details/address/#{station["code"]}.json", "w") { |f| f.write(JSON.pretty_generate(h)) }
+    File.open("src/details/address/#{station["code"]}.json", "w") { |f| f.write(JSON.pretty_generate(h)) }
     data = data["results"][0]
     puts "GeocodeAPI success."
   end
@@ -88,28 +87,34 @@ def get_address(station)
   station["address"] = address
 end
 
-class CSVTest < Minitest::Test
+class CSVTest < FormatTest
   def setup
     @stations = []
-    @station_map = {}
-    @station_code_set = Set.new
-    @station_name_set = Set.new
-    @id_set = IDSet.new
-    @pref_cnt = Array.new(48, 0)
-
     @lines = []
-    @line_name_set = Set.new
-    @line_code_set = Set.new
+    @id_set = IDSet.new
+
+    # read data from csv file, fields validation not checked
+    read_station()
+    read_line()
+    # fill in blank id values, if needed
+    check_id()
+
+    self.check_init()
+
+    # read line details from other files
+    # registaration of station-line is defined here
+    read_line_details()
   end
 
-  def test_csv
-    read_station()
-    check_prefecture_cnt()
-    read_line()
-    check_id()
+  def test_station()
+    # fill in black address and post-code, if needed
     check_address()
-    check_line_details()
-    check_register()
+
+    self.check_station(false)
+  end
+
+  def test_line()
+    self.check_line(false)
   end
 
   def teardown
@@ -120,7 +125,7 @@ class CSVTest < Minitest::Test
     puts "OK"
 
     print "Write to json files..."
-    File.open("solved/line.json", "w") do |f|
+    File.open("src/solved/line.json", "w") do |f|
       list = @lines.map do |line|
         line.delete_if do |key, value|
           value == nil || (key == "closed" && !value)
@@ -129,7 +134,7 @@ class CSVTest < Minitest::Test
       end
       f.write(format_json(list, flat: true))
     end
-    File.open("solved/station.json", "w") do |f|
+    File.open("src/solved/station.json", "w") do |f|
       list = @stations.map do |s|
         s.delete_if { |key, value| value == nil }
         s.delete("closed") if !s["closed"]
@@ -144,37 +149,20 @@ class CSVTest < Minitest::Test
   end
 
   def read_station
-    puts "read station info."
-    dup_original_name = Hash.new([])
-    csv_each_line("station.csv") do |fields|
+    csv_each_line("src/station.csv") do |fields|
       csv_err("col size != 15") if fields.length != 15
       code = fields["code"].to_i
-      csv_err("code duplicate") if !@station_code_set.add?(code)
       id = read_value(fields, "id")
-      if id
-        if id.match(/^[0-9a-f]{6}/)
-          csv_err("id duplicate") if !@id_set.add?(id)
-        else
-          csv_err("invalide id value")
-        end
-      end
       impl = read_boolean(fields, "impl")
 
       name = read_value(fields, "name")
       name_original = read_value(fields, "original_name")
       name_kana = read_value(fields, "name_kana")
-      # 駅名重複なし
-      csv_err("name duplicated") if !@station_name_set.add?(name)
       lat = fields["lat"].to_f
       lng = fields["lng"].to_f
       pref = fields["prefecture"].to_i
-      csv_err("invalid prefecture value") if pref < 1 || pref > 47
-      @pref_cnt[pref] += 1 if impl
       closed = read_boolean(fields, "closed")
       attr = read_value(fields, "attr")
-      csv_err("invalid attr value") if impl && attr != "eco" && attr != "heat" && attr != "cool" && attr != "unknown"
-      csv_err("invalid attr value (no-impl)") if !impl && attr
-      csv_err("closed <=> attr") if impl && closed != (attr == "unknown")
       postal_code = read_value(fields, "postal_code")
       address = read_value(fields, "address")
       station = {}
@@ -194,58 +182,21 @@ class CSVTest < Minitest::Test
       station["open_date"] = read_date(fields, "open_date")
       station["closed_date"] = read_date(fields, "closed_date")
 
+      puts "Warning > may be invalid suffix of name:#{name_original}" if name_original.end_with?("駅", "停留所", "乗降場")
       # 登録路線用
       station["lines"] = []
       @stations << station
-      @station_map[code] = station
-      @station_map[name] = station
-
-      if name != name_original
-        dup_original_name[name_original] = [station, *dup_original_name[name_original]]
-      end
     end
-
-    # check name duplication
-    dup_original_name.each do |key, value|
-      if s = @station_map[key]
-        assert s["impl"] && value.select { |v| v["impl"] }.length == 0, "original_name '#{key}' duplicated, but no suffix in name of #{JSON.dump(s)}"
-        # may be value.length == 1
-        next
-      end
-      assert value.length > 1, "original_name '#{key}' not duplicated #{JSON.dump(value[0])}"
-    end
-
     impl_size = @stations.select { |s| s["impl"] }.length
     puts "station size: #{@stations.length} (impl #{impl_size})"
   end
 
-  def check_prefecture_cnt
-    print "check impl station size in each prefecture..."
-    csv_each_line("check/prefecture.csv") do |fields|
-      code = fields["code"].to_i
-      name = fields["name"]
-      size = fields["size"].to_i
-      assert_equal size, @pref_cnt[code], "station size(impl) mismatch"
-    end
-    puts "OK"
-  end
-
   def read_line
-    puts "read line info."
-    csv_each_line("line.csv") do |fields|
+    csv_each_line("src/line.csv") do |fields|
       csv_err("fields size != 12") if fields.length != 12
       code = fields["code"].to_i
-      csv_err("line code duplicated") if !@line_code_set.add?(code)
       id = read_value(fields, "id")
-      if id
-        if id.match(/^[0-9a-f]{6}/)
-          csv_err("id duplicate") if !@id_set.add?(id)
-        else
-          csv_err("invalid id value")
-        end
-      end
       name = read_value(fields, "name")
-      csv_err("line name duplicated.") if !@line_name_set.add?(name)
       name_kana = read_value(fields, "name_kana")
       name_formal = read_value(fields, "name_formal")
       station_size = fields["station_size"].to_i
@@ -257,7 +208,6 @@ class CSVTest < Minitest::Test
       impl = read_boolean(fields, "impl")
       closed_date = read_date(fields, "closed_date")
       puts "Warning > line closed date not defined #{name}" if closed && !closed_date
-      csv_err("line not closed, but date defined") if !closed && closed_date
 
       line = {}
       line["code"] = code
@@ -290,51 +240,33 @@ class CSVTest < Minitest::Test
       end
     end
     if write_id
-      write_csv("station.csv", STATION_FIELD, @stations)
-      write_csv("line.csv", LINE_FIELD, @lines)
+      write_csv("src/station.csv", STATION_FIELD, @stations)
+      write_csv("src/line.csv", LINE_FIELD, @lines)
       puts "id added and saved."
     end
   end
 
   def check_address
-    puts "add post&address value if needed."
     @stations.each do |station|
       write = false
       if !station["postal_code"] || !station["address"]
         get_address(station)
         write = true
       end
-      assert station["postal_code"].match(/[0-9]{3}-[0-9]{4}/), "invalide post code: #{JSON.dump(station)}"
+      assert station["postal_code"].match(PATTERN_POST), "invalide post code: #{JSON.dump(station)}"
       write_csv("station.csv", STATION_FIELD, @stations) if write
     end
   end
 
-  def check_line_details
-    print "check line details: station-list, polyline..."
-    @lines.each do |line|
-      # 路線の詳細情報
-      path = "details/line/#{line["code"]}.json"
-      assert File.exists?(path), "file:#{path} not fount. line:#{JSON.dump(line)}"
-      # 路線ポリラインは廃線,no-implのみ欠損許す
-      path = "polyline/solved/#{line["code"]}.json"
-      assert File.exists?(path) || line["closed"] || !line["impl"], "polyline not found. line:#{JSON.dump(line)}"
-    end
-    puts "OK"
-  end
+  def read_line_details
+    puts "reading line details..."
 
-  def check_register
-    puts "checking station-list"
-
-    line_impl_size = {}
-    csv_each_line("check/line.csv") do |fields|
-      name = fields["name"]
-      size = fields["size"].to_i
-      line_impl_size[name] = size
-    end
     @register = []
     @register << ["station_code", "line_code", "index", "numbering"]
     @lines.each do |line|
-      path = "details/line/#{line["code"]}.json"
+      # 路線の登録駅情報
+      path = "src/details/line/#{line["code"]}.json"
+      assert File.exists?(path), "file:#{path} not found. line:#{JSON.dump(line)}"
       details = read_json(path)
       assert_equal line["name"], details["name"], "name mismatch(details). file:#{line["code"]}.json line:#{JSON.dump(line)}"
       # 登録駅数の確認
@@ -345,7 +277,7 @@ class CSVTest < Minitest::Test
       impl_size = 0
 
       write = false
-      details["station_list"].map!.each_with_index do |s, i|
+      line["station_list"] = details["station_list"].map.each_with_index do |s, i|
         station_code = s["code"]
         station_name = s["name"]
         # 名前解決
@@ -372,40 +304,23 @@ class CSVTest < Minitest::Test
         station["lines"] << line["code"]
         index = i + 1
         # 駅ナンバリングを文字列表現
-        numbering = "NULL"
-        if s.key?("numbering")
-          numbering = s["numbering"].map do |n|
-            value = ""
-            if n.key?("symbol")
-              value << n["symbol"]
-            elsif symbol
-              value << symbol
-            end
-            value << n["index"]
-            next value
-          end.join("/")
-        end
+        numbering = format_numbering(s, symbol)
+        numbering = "NULL" if numbering == nil
         @register << [station_code, line_code, index, numbering]
         next sort_hash(s)
       end
 
-      # 更新あるなら路線詳細へ反映
+      # 更新あるなら駅登録詳細へ反映
       if write
         File.open(path, "w:utf-8") do |f|
           f.write(format_json(details, flat_array: ["station_list"]))
         end
       end
-      # 登録駅数の再度確認
-      if line["impl"]
-        size = line_impl_size[line["name"]]
-        assert size, "line:#{line["name"]} not found at check/line. at station_list #{JSON.dump(line)}"
-        assert_equal size, impl_size, "station size(impl) mismatch at station_list #{JSON.dump(line)}"
-      end
-    end
-    # 路線登録されているか
-    @stations.each do |station|
-      station["lines"].uniq!
-      assert station["lines"].length > 0, "station not registered in any line. #{JSON.dump(station)}"
+
+      # 路線ポリラインは廃線,no-implのみ欠損許す
+      path = "src/polyline/solved/#{line["code"]}.json"
+      assert File.exists?(path) || line["closed"] || !line["impl"], "polyline not found. line:#{JSON.dump(line)}"
+      # no validation
     end
   end
 end
