+++ src/solved/station.json
@@ -1627,7 +1627,7 @@
   {"code":1123004,"id":"127021","name":"久ノ浜","name_kana":"ひさのはま","lat":37.142465,"lng":140.995741,"prefecture":7,"lines":[11230],"attr":"cool","postal_code":"979-0333","address":"いわき市久之浜町久之浜字北荒蒔","impl":true},
   {"code":1123005,"id":"3d3c97","name":"末続","name_kana":"すえつぎ","lat":37.172156,"lng":140.995565,"prefecture":7,"lines":[11230],"attr":"cool","postal_code":"979-0331","address":"いわき市久之浜町大字末続","impl":true},
   {"code":1123006,"id":"e42cc8","name":"広野(福島)","original_name":"広野","name_kana":"ひろの","lat":37.21157,"lng":140.998769,"prefecture":7,"lines":[11230],"attr":"cool","postal_code":"979-0403","address":"双葉郡広野町大字下浅見川","impl":true},
-  {"code":1123035,"id":"4ea041","name":"Jヴィレッジ","name_kana":"じぇいゔぃれっじ","lat":37.242583,"lng":141.0075,"prefecture":7,"lines":[11230],"attr":"cool","postal_code":"979-0513","address":"双葉郡楢葉町大字山田岡字下岩沢1番地2","open_date":"2019-04-20","impl":true},
+  {"code":1123035,"id":"4ea041","name":"Jヴィレッジ","name_kana":"じぇいゔぃれっじ","lat":37.242583,"lng":141.0075,"prefecture":7,"lines":[11230],"attr":"cool","postal_code":"979-0513","address":"双葉郡楢葉町大字山田岡字下岩沢1番地2","open_date":"2019-04-20","impl":true},
   {"code":1123007,"id":"650492","name":"木戸","name_kana":"きど","lat":37.257875,"lng":141.00247,"prefecture":7,"lines":[11230],"attr":"cool","postal_code":"979-0513","address":"双葉郡楢葉町大字山田岡","impl":true},
   {"code":1123008,"id":"a91215","name":"竜田","name_kana":"たつた","lat":37.282485,"lng":141.002822,"prefecture":7,"lines":[11230],"attr":"cool","postal_code":"979-0603","address":"双葉郡楢葉町大字井出","impl":true},
   {"code":1123009,"id":"6571d6","name":"富岡","name_kana":"とみおか","lat":37.335264,"lng":141.022274,"prefecture":7,"lines":[11230],"attr":"cool","postal_code":"979-1121","address":"双葉郡富岡町大字仏浜","impl":true},
@@ -9789,7 +9789,7 @@
   {"code":9911606,"id":"fa2999","name":"中の丘","name_kana":"なかのおか","closed":true,"lat":43.474583,"lng":142.169806,"prefecture":1,"lines":[99116],"postal_code":"075-0167","address":"北海道芦別市中の丘町","open_date":"1949-09-15","closed_date":"1989-03-26","impl":false},
   {"code":9911607,"id":"874de4","name":"幸町","name_kana":"さいわいちょう","closed":true,"lat":43.467,"lng":142.170889,"prefecture":1,"lines":[99116],"postal_code":"075-0163","address":"北海道芦別市緑泉町１９","open_date":"1958-01-20","closed_date":"1989-03-26","impl":false},
   {"code":9911608,"id":"ba8047","name":"緑泉","name_kana":"ろくせん","closed":true,"lat":43.461583,"lng":142.165889,"prefecture":1,"lines":[99116],"postal_code":"075-0163","address":"北海道芦別市緑泉町１６−６","open_date":"1945-12-15","closed_date":"1989-03-26","impl":false},
-  {"code":9911609,"id":"0a5cf7","name":"西町アパート前","name_kana":"にしまちアパートまえ","closed":true,"lat":43.457417,"lng":142.163833,"prefecture":1,"lines":[99116],"postal_code":"075-0164","address":"北海道芦別市頼城町４−１０２","open_date":"1958-01-20","closed_date":"1989-03-26","impl":false},
+  {"code":9911609,"id":"0a5cf7","name":"西町アパート前","name_kana":"にしまちあぱーとまえ","closed":true,"lat":43.457417,"lng":142.163833,"prefecture":1,"lines":[99116],"postal_code":"075-0164","address":"北海道芦別市頼城町４−１０２","open_date":"1958-01-20","closed_date":"1989-03-26","impl":false},
   {"code":9911610,"id":"4bbe3d","name":"芦の湯前","name_kana":"あしのゆまえ","closed":true,"lat":43.453111,"lng":142.158,"prefecture":1,"lines":[99116],"postal_code":"075-0164","address":"北海道芦別市頼城町４−１２","open_date":"1958-01-20","closed_date":"1989-03-26","impl":false},
   {"code":9911611,"id":"cf33fb","name":"頼城","name_kana":"らいじょう","closed":true,"lat":43.449361,"lng":142.158056,"prefecture":1,"lines":[99116],"postal_code":"075-0164","address":"北海道芦別市頼城町６９","open_date":"1945-12-15","closed_date":"1989-03-26","impl":false},
   {"code":9911612,"id":"846ef2","name":"玉川(北海道)","original_name":"玉川","name_kana":"たまがわ","closed":true,"lat":43.442056,"lng":142.157778,"prefecture":1,"lines":[99116],"postal_code":"075-0164","address":"北海道芦別市頼城町１００","closed_date":"1989-03-26","impl":false},
@@ -9851,7 +9851,7 @@
   {"code":9922513,"id":"76ac4e","name":"尾松","name_kana":"おまつ","closed":true,"lat":38.819361,"lng":140.975528,"prefecture":4,"lines":[99225],"postal_code":"989-5341","address":"宮城県栗原市栗駒稲屋敷上ノ山１７","open_date":"1942-12-01","closed_date":"2007-04-01","impl":false},
   {"code":9922514,"id":"da7731","name":"鶯沢","name_kana":"うぐいすざわ","closed":true,"lat":38.826667,"lng":140.949028,"prefecture":4,"lines":[99225],"postal_code":"989-5403","address":"宮城県栗原市鶯沢北郷寺原６８","open_date":"1942-12-01","closed_date":"2007-04-01","impl":false},
   {"code":9922515,"id":"a87041","name":"鶯沢工業高校前","name_kana":"うぐいすざわこうぎょうこうこうまえ","closed":true,"lat":38.807111,"lng":140.925528,"prefecture":4,"lines":[99225],"postal_code":"989-5402","address":"宮城県栗原市鶯沢南郷上新反田２７","open_date":"1952-04-01","closed_date":"2007-04-01","impl":false},
-  {"code":9922516,"id":"daaa6c","name":"細倉マインパーク前","name_kana":"ほそくらマインパークまえ","closed":true,"lat":38.803028,"lng":140.906028,"prefecture":4,"lines":[99225],"postal_code":"989-5402","address":"宮城県栗原市鶯沢南郷原１２","open_date":"1942-12-01","closed_date":"2007-04-01","impl":false},
+  {"code":9922516,"id":"daaa6c","name":"細倉マインパーク前","name_kana":"ほそくらまいんぱーくまえ","closed":true,"lat":38.803028,"lng":140.906028,"prefecture":4,"lines":[99225],"postal_code":"989-5402","address":"宮城県栗原市鶯沢南郷原１２","open_date":"1942-12-01","closed_date":"2007-04-01","impl":false},
   {"code":9922601,"id":"857cb4","name":"東八幡平","name_kana":"ひがしはちまんたい","closed":true,"lat":39.920409,"lng":140.974573,"prefecture":3,"lines":[99226],"postal_code":"028-7303","address":"岩手県八幡平市柏台１丁目２-２８","open_date":"1948-03-15","closed_date":"1972-10-10","impl":false},
   {"code":9922602,"id":"6e621a","name":"鹿野","name_kana":"ししの","closed":true,"lat":39.922228,"lng":141.047909,"prefecture":3,"lines":[99226],"postal_code":"028-7302","address":"岩手県八幡平市松尾寄木第２９地割４０−９","open_date":"1948-03-15","closed_date":"1972-10-10","impl":false},
   {"code":9922603,"id":"0e3efc","name":"田頭","name_kana":"でんどう","closed":true,"lat":39.922303,"lng":141.077675,"prefecture":3,"lines":[99226],"postal_code":"028-7112","address":"岩手県八幡平市田頭第２６地割２７","closed_date":"1972-10-10","impl":false},
