+++ src/station.csv
@@ -1627,7 +1627,7 @@ code,id,name,original_name,name_kana,lat,lng,prefecture,postal_code,address,clos
 1123004,127021,久ノ浜,久ノ浜,ひさのはま,37.142465,140.995741,7,979-0333,いわき市久之浜町久之浜字北荒蒔,0,NULL,NULL,1,cool
 1123005,3d3c97,末続,末続,すえつぎ,37.172156,140.995565,7,979-0331,いわき市久之浜町大字末続,0,NULL,NULL,1,cool
 1123006,e42cc8,広野(福島),広野,ひろの,37.21157,140.998769,7,979-0403,双葉郡広野町大字下浅見川,0,NULL,NULL,1,cool
-1123035,4ea041,Jヴィレッジ,Jヴィレッジ,じぇいゔぃれっじ,37.242583,141.0075,7,979-0513,双葉郡楢葉町大字山田岡字下岩沢1番地2,0,2019-04-20,NULL,1,cool
+1123035,4ea041,Jヴィレッジ,Jヴィレッジ,じぇいゔぃれっじ,37.242583,141.0075,7,979-0513,双葉郡楢葉町大字山田岡字下岩沢1番地2,0,2019-04-20,NULL,1,cool
 1123007,650492,木戸,木戸,きど,37.257875,141.00247,7,979-0513,双葉郡楢葉町大字山田岡,0,NULL,NULL,1,cool
 1123008,a91215,竜田,竜田,たつた,37.282485,141.002822,7,979-0603,双葉郡楢葉町大字井出,0,NULL,NULL,1,cool
 1123009,6571d6,富岡,富岡,とみおか,37.335264,141.022274,7,979-1121,双葉郡富岡町大字仏浜,0,NULL,NULL,1,cool
@@ -9789,7 +9789,7 @@ code,id,name,original_name,name_kana,lat,lng,prefecture,postal_code,address,clos
 9911606,fa2999,中の丘,中の丘,なかのおか,43.474583,142.169806,1,075-0167,北海道芦別市中の丘町,1,1949-09-15,1989-03-26,0,NULL
 9911607,874de4,幸町,幸町,さいわいちょう,43.467,142.170889,1,075-0163,北海道芦別市緑泉町１９,1,1958-01-20,1989-03-26,0,NULL
 9911608,ba8047,緑泉,緑泉,ろくせん,43.461583,142.165889,1,075-0163,北海道芦別市緑泉町１６−６,1,1945-12-15,1989-03-26,0,NULL
-9911609,0a5cf7,西町アパート前,西町アパート前,にしまちアパートまえ,43.457417,142.163833,1,075-0164,北海道芦別市頼城町４−１０２,1,1958-01-20,1989-03-26,0,NULL
+9911609,0a5cf7,西町アパート前,西町アパート前,にしまちあぱーとまえ,43.457417,142.163833,1,075-0164,北海道芦別市頼城町４−１０２,1,1958-01-20,1989-03-26,0,NULL
 9911610,4bbe3d,芦の湯前,芦の湯前,あしのゆまえ,43.453111,142.158,1,075-0164,北海道芦別市頼城町４−１２,1,1958-01-20,1989-03-26,0,NULL
 9911611,cf33fb,頼城,頼城,らいじょう,43.449361,142.158056,1,075-0164,北海道芦別市頼城町６９,1,1945-12-15,1989-03-26,0,NULL
 9911612,846ef2,玉川(北海道),玉川,たまがわ,43.442056,142.157778,1,075-0164,北海道芦別市頼城町１００,1,NULL,1989-03-26,0,NULL
@@ -9851,7 +9851,7 @@ code,id,name,original_name,name_kana,lat,lng,prefecture,postal_code,address,clos
 9922513,76ac4e,尾松,尾松,おまつ,38.819361,140.975528,4,989-5341,宮城県栗原市栗駒稲屋敷上ノ山１７,1,1942-12-01,2007-04-01,0,NULL
 9922514,da7731,鶯沢,鶯沢,うぐいすざわ,38.826667,140.949028,4,989-5403,宮城県栗原市鶯沢北郷寺原６８,1,1942-12-01,2007-04-01,0,NULL
 9922515,a87041,鶯沢工業高校前,鶯沢工業高校前,うぐいすざわこうぎょうこうこうまえ,38.807111,140.925528,4,989-5402,宮城県栗原市鶯沢南郷上新反田２７,1,1952-04-01,2007-04-01,0,NULL
-9922516,daaa6c,細倉マインパーク前,細倉マインパーク前,ほそくらマインパークまえ,38.803028,140.906028,4,989-5402,宮城県栗原市鶯沢南郷原１２,1,1942-12-01,2007-04-01,0,NULL
+9922516,daaa6c,細倉マインパーク前,細倉マインパーク前,ほそくらまいんぱーくまえ,38.803028,140.906028,4,989-5402,宮城県栗原市鶯沢南郷原１２,1,1942-12-01,2007-04-01,0,NULL
 9922601,857cb4,東八幡平,東八幡平,ひがしはちまんたい,39.920409,140.974573,3,028-7303,岩手県八幡平市柏台１丁目２-２８,1,1948-03-15,1972-10-10,0,NULL
 9922602,6e621a,鹿野,鹿野,ししの,39.922228,141.047909,3,028-7302,岩手県八幡平市松尾寄木第２９地割４０−９,1,1948-03-15,1972-10-10,0,NULL
 9922603,0e3efc,田頭,田頭,でんどう,39.922303,141.077675,3,028-7112,岩手県八幡平市田頭第２６地割２７,1,NULL,1972-10-10,0,NULL
